requirements.txt
python-dotenv
google-api-python-client
google-cloud-datastore
google-cloud-resource-manager
google-cloud-compute
google-cloud-securitycenter
google-cloud-org-policy
google-cloud-securitycentermanagement
google-cloud-network-management
google-cloud-dns
celery
fastapi
protobuf==3.20.3
